fix(R-11): extract DividendCheckView and AssetPriceView UserControls from MainWindow
MainWindow was a God Object handling three unrelated feature areas. Each tab's
XAML content and event handlers are now in dedicated UserControls injected via DI,
leaving MainWindow as a thin shell responsible only for wiring up the navigation tree.
